chromium*: set by default debugging symbol level to 1
Instead of forcing symbol level to 0, that removes all debugging
information from build, set the default level to 1.

This change adds two new PACKAGECONFIG named "debug" (for setting
the symbol level for the whole build to 2) and "debug-blink" (for
setting the symbol level for blink part to 2). By default, if not
present, symbol level will be set to 1.

Symbol level is equivalent to the -g level (symbol_level=1 means
-g1, symbol_level=2 means -g or -g2).

This is copied from the webOS OSE web runtime Yocto recipe.

Bug-AGL: SPEC-4233
Change-Id: I70a983a04497aefdf9fde7074bbf0cfdf0bb8eab