chromium, wam: update to chromium87 tree.
Update to chromium87 tree, and wam adapted to chromium87. This does
not include yet the webOS OSE QtLess version, but the submission
right before QtLess work. This is because in webOS OSE current
chromium87 release, the only work after the submission we are using
is QtLess, without adding any new functionality.

Renamed the recipe and package name to be "chromium" instead of
"chromium84" or "chromium87". Now we use PV to state the version of
the upstream Chromium.

The Chromium87 tree history has been heavily simplified to make it
easier to rebase in next upgrades.

Bug-AGL: SPEC-4142, SPEC-4225, SPEC-4227, SPEC-4228

Change-Id: I6caed396f758dd80291803266f6a59f1fec406c4